+++ nuxeo-drive-client/nxdrive/__init__.py
@@ -25,4 +25,4 @@ To declare a beta, use this schema:
 """
 
 __author__ = 'Nuxeo'
-__version__ = '2.5.4'
+__version__ = '2.5.5'